need to figure out a way to override session ID for checkin
