+++ controllers/index.js
@@ -1,10 +1,12 @@
 const router = require("express").Router();
-const homeRoutes = require("./home-routes.js");
 
+const homeRoutes = require("./home-routes.js");
+const dashboardRoutes = require("./dashboard-routes");
 const apiRoutes = require("./api");
 
 router.use("/api", apiRoutes);
 router.use("/", homeRoutes);
+router.use("/dash", dashboardRoutes);
 
 router.use((req, res) => {
   res.status(404).end();
