+++ views/dash.handlebars
@@ -22,3 +22,4 @@
                 </div>
             </div>
         </div>
+        <script src="/js/checkin.js"></script>
