+++ views/homepage.handlebars
@@ -5,7 +5,7 @@
 				<div class="hero-body">
 					<div class="container">
 						<h1 class="title">
-							Hello!
+							Hello, {{Student.name}}!
 						</h1>
 						<h2 class="subtitle">
 							Click on a category to narrow down your search.
